move keyslots to separate file, and rename them
