+++ crates/crypto/src/header/mod.rs
@@ -3,3 +3,4 @@
 
 pub mod file;
 pub mod serialization;
+pub mod keyslot;
